Extract item.Catalog as item.InMemoryCatalog
Collections of a type go within the type's package; but I put it in a
separate file because it's for testing; and I renamed it both to make
the concrete type self-explanatory and avoid confusion with the
implemented interface, even if it's in another package.
Also, test fixtures go within the data they model for cohesion; they
should not litter the prod-release because I think the compiler purches
dead code after removing test-files?